pyproject.toml
[tool.poetry]
name = "elleelleaime"
version = "0.0.1"
description = "Automatic Program Repair with Large Language Models"
license = "MIT"
authors = [
    "André Silva <andreans@kth.se>"
]
repository = "https://github.com/ASSERT-KTH/elle-elle-aime"
homepage = "https://github.com/ASSERT-KTH/elle-elle-aime"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
openai = "^1.35.1"
python-dotenv = "^1.0.1"
transformers = {extras = ["torch"], version = "^4.33.0"}
fire = "^0.6.0"
tqdm = "^4.66.4"
numpy = "^1.26.4"
unidiff = "^0.7.5"
backoff = "^2.2.1"
black = "^24.4.2"
torch = "^2.3.1"
accelerate = "^0.32.1"
sentencepiece = "^0.2.0"
datasets = "^2.20.0"
peft = "^0.11.1"
bitsandbytes = "^0.43.1"
evaluate = "^0.4.2"
safetensors = "^0.4.3"
google-generativeai = "^0.7.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.2.0"
